Update index.md
Just playing around trying to get used to github really (and the difference between a branch, a commit and a pull request - I think I get it now).

This just provides an outline of what could sit on a home page which is open for comment. I'd like to see how comments works once I make the following pull request.